+++ cql3/restrictions/restriction.hh
@@ -48,7 +48,7 @@
 #include <variant>
 #include <vector>
 
-#include <fmt/ostream.h>
+#include <fmt/core.h>
 #include <seastar/core/shared_ptr.hh>
 #include <seastar/core/sstring.hh>
 #include "utils/overloaded_functor.hh"
@@ -252,31 +252,7 @@ public:
 }
 
 /// Required for fmt::join() to work on expression.
-template <>
-struct fmt::formatter<cql3::restrictions::expression> {
-    constexpr auto parse(format_parse_context& ctx) {
-        return ctx.end();
-    }
-
-    template <typename FormatContext>
-    auto format(const cql3::restrictions::expression& expr, FormatContext& ctx) {
-        std::ostringstream os;
-        os << expr;
-        return format_to(ctx.out(), "{}", os.str());
-    }
-};
+template <> struct fmt::formatter<cql3::restrictions::expression>;
 
 /// Required for fmt::join() to work on column_value.
-template <>
-struct fmt::formatter<cql3::restrictions::column_value> {
-    constexpr auto parse(format_parse_context& ctx) {
-        return ctx.end();
-    }
-
-    template <typename FormatContext>
-    auto format(const cql3::restrictions::column_value& col, FormatContext& ctx) {
-        std::ostringstream os;
-        os << col;
-        return format_to(ctx.out(), "{}", os.str());
-    }
-};
+template <> struct fmt::formatter<cql3::restrictions::column_value>;
